requirements.txt
wheel
dialog_bot_sdk
python-dotenv
grpcio
